chore: add a new password feature
